# src/flask/openapisample.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Full Stack FastAPI Project",
    "description": "Sample API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/payments/create-intent": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Payment Intent Endpoint",
        "description": "Create a Stripe PaymentIntent for a one-time donation.",
        "operationId": "create_payment_intent_endpoint_api_v1_payments_create_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/payments/create-subscription": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Subscription Endpoint",
        "description": "Create a Stripe Checkout Session for a recurring donation.",
        "operationId": "create_subscription_endpoint_api_v1_payments_create_subscription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/payments/": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Get User Payments",
        "description": "Get authenticated user's payment history.",
        "operationId": "get_user_payments_api_v1_payments__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{payment_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Get Payment",
        "description": "Get a single payment detail.",
        "operationId": "get_payment_api_v1_payments__payment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutSessionResponse": {
        "properties": {
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "checkout"
          },
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "client_secret",
          "checkout_url"
        ],
        "title": "CheckoutSessionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentCreate": {
        "properties": {
          "amount_cents": {
            "type": "integer",
            "maximum": 999999999,
            "exclusiveMinimum": 0,
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "title": "Currency",
            "default": "usd"
          },
          "frequency": {
            "type": "string",
            "pattern": "^(one_time|recurring)$",
            "title": "Frequency"
          },
          "donor_email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Donor Email"
          },
          "donor_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Donor Name"
          },
          "metadata_json": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          }
        },
        "type": "object",
        "required": [
          "amount_cents",
          "frequency"
        ],
        "title": "PaymentCreate"
      },
      "PaymentIntentResponse": {
        "properties": {
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          },
          "payment_intent_id": {
            "type": "string",
            "title": "Payment Intent Id"
          }
        },
        "type": "object",
        "required": [
          "client_secret",
          "payment_intent_id"
        ],
        "title": "PaymentIntentResponse"
      },
      "PaymentPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stripe_payment_intent_id": {
            "type": "string",
            "title": "Stripe Payment Intent Id"
          },
          "stripe_subscription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Subscription Id"
          },
          "donor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Donor Email"
          },
          "donor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Donor Name"
          },
          "receipt_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Url"
          },
          "created_on": {
            "type": "string",
            "format": "date-time",
            "title": "Created On"
          },
          "updated_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated On"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount_cents",
          "currency",
          "status",
          "stripe_payment_intent_id",
          "stripe_subscription_id",
          "donor_email",
          "donor_name",
          "receipt_url",
          "created_on",
          "updated_on"
        ],
        "title": "PaymentPublic"
      },
      "PaymentsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PaymentPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "PaymentsPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}